refactor(ocean-road): :building_construction: modified semantics and colors with tokens
