+++ apps/blog-client/ui/tag-item/tag-item.styled.tsx
@@ -1,11 +1,10 @@
-import variables from '@coldsurfers/design-tokens/dist/js/color/semantic/variables'
-import { media } from '@coldsurfers/ocean-road'
+import { media, semantics } from '@coldsurfers/ocean-road'
 import { css } from '@emotion/react'
 import styled from '@emotion/styled'
 import { Paragraph } from '../paragraph'
 
 export const StyledTagItem = styled.div`
-  background-color: ${variables.color.background[4]};
+  background-color: ${semantics.color.background[4]};
   opacity: 0.8;
   padding: 0.25rem 0.5rem;
   border-radius: 0.25rem;
@@ -14,7 +13,7 @@ export const StyledTagItem = styled.div`
   margin-right: 0.5rem;
   margin-bottom: 0.5rem;
 
-  color: ${variables.color.foreground[3]};
+  color: ${semantics.color.foreground[3]};
 `
 
 export const StyledTagItemText = styled(Paragraph)`
