+++ packages/ocean-road/src/index.ts
@@ -6,4 +6,5 @@ export * from './spinner'
 export * from './text'
 export * from './text-input'
 export * from './toast'
+export * from './tokens'
 export * from './utils'
