+++ packages/ocean-road/src/spinner/spinner.tsx
@@ -1,5 +1,5 @@
-import variables from '@coldsurfers/design-tokens/dist/js/color/semantic/variables'
 import FadeLoader from 'react-spinners/FadeLoader'
+import { semantics } from '../tokens'
 import { StyledPageLoadingWrapper } from './spinner.styled'
 import { SpinnerVariant } from './spinner.types'
 
@@ -9,7 +9,7 @@ export const Spinner = ({ variant }: { variant: SpinnerVariant }) => {
       return (
         <StyledPageLoadingWrapper>
           <StyledPageLoadingWrapper>
-            <FadeLoader height="13" color={variables.color.foreground[1]} />
+            <FadeLoader height="13" color={semantics.color.foreground[1]} />
           </StyledPageLoadingWrapper>
         </StyledPageLoadingWrapper>
       )
