+++ packages/ocean-road/src/text/text.styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import { semantics } from '../contexts/ColorSchemeProvider'
+import { semantics } from '../tokens'
 
 export const StyledTextContainer = styled.span`
   color: ${semantics.color.foreground[1]};
