+++ packages/ocean-road/src/tokens/tokens.ts
@@ -1,0 +1,6 @@
+import colorDesignTokens from '@coldsurfers/design-tokens/dist/js/color/variables'
+import semanticDesignTokens from '@coldsurfers/design-tokens/dist/js/semantic/variables'
+import { ColorDesignTokens } from './tokens.types'
+
+export const colors: ColorDesignTokens = colorDesignTokens
+export const semantics = semanticDesignTokens
