+++ packages/ocean-road/src/tokens/tokens.types.ts
@@ -1,0 +1,7 @@
+import colorDesignTokens from '@coldsurfers/design-tokens/dist/js/color/variables'
+import darkColorDesignTokens from '@coldsurfers/design-tokens/dist/json/color/variables-dark.json'
+import lightColorDesignTokens from '@coldsurfers/design-tokens/dist/json/color/variables-light.json'
+
+export type ColorDesignTokens = typeof colorDesignTokens
+export type DarkColorDesignTokens = typeof darkColorDesignTokens
+export type LightColorDesignTokens = typeof lightColorDesignTokens
